docpro/backend/requirements.txt
# Core Django Framework
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0

# ASGI and WebSockets
channels>=4.0,<5.0
daphne>=4.0,<5.0
asgiref>=3.7,<4.0

# Background Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# PDF Processing
pypdf>=4.0,<5.0
pymupdf>=1.23,<2.0
pdfplumber>=0.10,<1.0

# Document Processing
python-docx>=1.0,<2.0
docx2pdf>=0.1.8,<1.0

# Image and OCR
opencv-python>=4.8,<5.0
pytesseract>=0.3,<1.0

# NLP and ML
spacy>=3.6,<4.0
numpy>=1.24,<2.0

# Web Scraping and HTML Parsing
beautifulsoup4>=4.12,<5.0

# Configuration and Utilities
python-dotenv>=1.0,<2.0
drf-yasg>=1.21,<2.0
django-cors-headers>=4.3,<5.0
